AP_EFI: EFI_Serial was trying to do a double calc

+++ libraries/AP_EFI/AP_EFI_Serial_MS.cpp
@@ -124,7 +124,7 @@ bool AP_EFI_Serial_MS::read_incoming_realtime_data()
             case TPS_MSB:
                 temp_float = (float)((data << 8) + read_byte_CRC32())/10.0f;
                 offset++;
-                internal_state.throttle_position_percent = round(temp_float);
+                internal_state.throttle_position_percent = roundf(temp_float);
                 break;
             case AFR1_MSB:
                 temp_float = (float)((data << 8) + read_byte_CRC32())/10.0f;
